requirements.txt
# AI PPT Generator - 依赖列表
# 安装: pip install -r requirements.txt

# Web 框架
Flask>=2.3.0
Flask-CORS>=4.0.0
flask-swagger-ui>=4.0.0

# 数据库
SQLAlchemy>=2.0.0
Flask-SQLAlchemy>=3.1.0

# 认证
PyJWT>=2.8.0
bcrypt>=4.1.0

# PPT 导出
python-pptx>=0.6.23

# 数据处理
pandas>=2.0.0

# 实用工具
python-dotenv>=1.0.0
marshmallow>=3.20.0
PyYAML>=6.0

# 测试
pytest>=7.4.0
pytest-cov>=4.1.0

# 开发工具
black>=23.0.0
flake8>=6.1.0
